requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.4
    # via
    #   black
    #   virtualenv
async-generator==1.10
    # via
    #   -r requirements/prod.txt
    #   httpx
attrs==21.2.0
    # via
    #   black
    #   pytest
black==19.10b0
    # via -r requirements/dev.in
certifi==2020.12.5
    # via
    #   -r requirements/prod.txt
    #   httpx
cfgv==3.3.0
    # via pre-commit
click==8.0.1
    # via
    #   black
    #   pip-tools
contextvars==2.4
    # via
    #   -r requirements/prod.txt
    #   sniffio
coverage[toml]==5.5
    # via
    #   -r requirements/dev.in
    #   pytest-cov
distlib==0.3.1
    # via virtualenv
filelock==3.0.12
    # via virtualenv
h11==0.12.0
    # via
    #   -r requirements/prod.txt
    #   httpcore
httpcore==0.13.3
    # via
    #   -r requirements/prod.txt
    #   httpx
httpx==0.18.1
    # via -r requirements/prod.txt
identify==2.2.6
    # via pre-commit
idna==3.1
    # via
    #   -r requirements/prod.txt
    #   rfc3986
immutables==0.15
    # via
    #   -r requirements/prod.txt
    #   contextvars
importlib-metadata==4.3.0
    # via
    #   click
    #   pep517
    #   pluggy
    #   pre-commit
    #   pytest
    #   virtualenv
importlib-resources==5.1.4
    # via
    #   pre-commit
    #   virtualenv
iniconfig==1.1.1
    # via pytest
nodeenv==1.6.0
    # via pre-commit
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements/dev.in
pluggy==0.13.1
    # via pytest
pre-commit==2.13.0
    # via -r requirements/dev.in
prometheus-client==0.10.1
    # via -r requirements/prod.txt
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest-asyncio==0.15.1
    # via -r requirements/dev.in
pytest-cov==2.12.0
    # via -r requirements/dev.in
pytest==6.2.4
    # via
    #   pytest-asyncio
    #   pytest-cov
python-decouple==3.4
    # via -r requirements/prod.txt
pyyaml==5.4.1
    # via pre-commit
regex==2021.4.4
    # via black
rfc3986[idna2008]==1.5.0
    # via
    #   -r requirements/prod.txt
    #   httpx
six==1.16.0
    # via virtualenv
sniffio==1.2.0
    # via
    #   -r requirements/prod.txt
    #   httpcore
    #   httpx
toml==0.10.2
    # via
    #   black
    #   coverage
    #   pep517
    #   pre-commit
    #   pytest
typed-ast==1.4.3
    # via black
typing-extensions==3.10.0.0
    # via importlib-metadata
virtualenv==20.4.7
    # via pre-commit
zipp==3.4.1
    # via
    #   importlib-metadata
    #   importlib-resources
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
